requirements.txt
selenium
pandas
streamlit
plotly
webdriver-manager
python-dotenv
beautifulsoup4
lxml